--- aspnetcore/fundamentals/primitives/change-tokens/sample/Data/AppDbContext.cs
@@ -1,14 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-
-namespace ChangeTokenSample.Data
-{
-    public class AppDbContext : DbContext
-    {
-        public AppDbContext(DbContextOptions options)
-            : base(options)
-        {
-        }
-
-        public DbSet<Message> Messages { get; set; }
-    }
-}
