--- aspnetcore/fundamentals/primitives/change-tokens/sample/Data/Message.cs
@@ -1,8 +1,0 @@
-namespace ChangeTokenSample.Data
-{
-    public class Message
-    {
-        public int Id { get; set; }
-        public string Text { get; set; }
-    }
-}
