--- aspnetcore/fundamentals/primitives/change-tokens/sample/Enums/Enums.cs
@@ -1,20 +1,0 @@
-namespace ChangeTokenSample.Enums
-{
-    #region snippet1
-    public enum FileState
-    { 
-        NotUpdated,
-        Updated
-    }
-    #endregion
-
-    #region snippet2
-    public enum MessageChangeType
-    {
-        None,
-        Add,
-        Delete,
-        DeleteAll
-    }
-    #endregion
-}
