--- aspnetcore/fundamentals/primitives/change-tokens/sample/Pages/FileWatching.cshtml
@@ -1,46 +1,0 @@
-﻿@page
-@model FileWatchingModel
-@{
-    ViewData["Title"] = "File changes";
-}
-
-<h1>@ViewData["Title"]</h1>
-
-<h2>Watching for changes in a file</h2>
-
-<div class="row">
-    <div class="col-md-8">
-
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">File status</h3>
-            </div>
-            <div class="panel-body">
-                <ol>
-                    <li>Open file <em>@Model.FileName</em> (<code>@Model.FilePath</code>) in a text editor and modify the file's contents.</li>
-                    <li>Save the file.</li>
-                    <li>The console logs a message when the file watcher triggers a callback on the change token.<img width="290" height="73" alt="Console showing a message that the file Spirits of the Dead changed on 10/27/2017 at 9:21:27 AM" src="~/images/console-file-change.png"></li>
-                    <li>Click the <b>Get Current Status</b> button to update the page with the file's content and latest modification status.</li>
-                    <li>Click the <b>Reset</b> button to change the file's status back to <span class="label label-danger">NotUpdated</span>.</li>
-                </ol>
-                <p><b>File: <i>@Model.FileName</i></b></p>
-                <textarea disabled style="resize:none;width:100%;height:100px;overflow-y:auto">@Model.FileContents</textarea>
-                <p><b>File State:</b> <span class="label label-@ViewData["UpdatedStyle"]">@Model.CurrentFileState</span></p>
-                <p><a href="/FileWatching" class="btn btn-default" role="button">Get Current Status</a></p>
-            </div>
-        </div>
-
-        <div class="panel panel-danger">
-            <div class="panel-heading">
-                <h3 class="panel-title">Reset</h3>
-            </div>
-            <div class="panel-body">
-                <p>Clicking this button resets the file state to <span class="label label-danger">NotUpdated</span>.</p>
-                <form method="post">
-                    <button type="submit" asp-method="post" asp-page-handler="ResetFileState" class="btn btn-danger">Reset</button>
-                </form>
-            </div>
-        </div>
-
-    </div>
-</div>
