+++ aspnetcore/fundamentals/primitives/change-tokens/sample/Pages/Index.cshtml
@@ -9,15 +9,27 @@
 <h2>Detect changes with Change Token</h2>
 
 <div class="row">
-    <div class="col-md-4">
+    <div class="col-md-8">
+
         <div class="panel panel-default">
             <div class="panel-heading">
-                <h3 class="panel-title">Examples</h3>
+                <h3 class="panel-title">Logging levels</h3>
             </div>
             <div class="panel-body">
-                <p><a href="/ConfigWatching" class="btn btn-primary" role="button">Watching for configuration changes</a></p>
-                <p><a href="/FileWatching" class="btn btn-primary" role="button"">Watching for changes in a file</a></p>
-                <p><a href="/ObjectWatching" class="btn btn-primary" role="button">Watching for changes in an object</a></p>
+                <ol>
+                    <li>Open the <em>appsettings.Production.json</em> file (or <em>appsettings.Development.json</em> if running the sample in the Development environment) in a text editor and set the logging levels to any of these valid values: <span class="label label-default">Trace</span>, <span class="label label-success">Debug</span>, <span class="label label-info">Information</span>, <span class="label label-warning">Warning</span>, <span class="label label-primary">Error</span>, or <span class="label label-danger">Critical</span>.</li>
+                    <li>Save the file.</li>
+                    <li>The console logs a message when the file watcher triggers a callback on the change token.<img width="258" height="70" alt="Console showing a message that the configuration changed on 10/27/2017 at 9:30:17 AM" src="~/images/console-config-change.png"></li>
+                    <li>Click the <b>Get Current Values</b> button to update the page with the latest values loaded into configuration.</li>
+                </ol>
+                <p>Include Scopes: <span class="label label-@ViewData["IncludeScopesStyle"]">@Model.IncludeScopes</span></p>
+                LogLevel
+                <ul>
+                    <li>Default: <span class="label label-@ViewData["DefaultLogLevelStyle"]">@Model.DefaultLogLevel</span></li>
+                    <li>System: <span class="label label-@ViewData["SystemLogLevelStyle"]">@Model.SystemLogLevel</span></li>
+                    <li>Microsoft: <span class="label label-@ViewData["MicrosoftLogLevelStyle"]">@Model.MicrosoftLogLevel</span></li>
+                </ul> 
+                <p><a asp-page="/Index" class="btn btn-default" role="button">Get Current Values</a></p>
             </div>
         </div>
