--- aspnetcore/fundamentals/primitives/change-tokens/sample/Pages/ObjectWatching.cshtml
@@ -1,77 +1,0 @@
-﻿@page
-@model ObjectWatchingModel
-@{
-    ViewData["Title"] = "Object changes";
-}
-
-<h1>@ViewData["Title"]</h1>
-
-<h2>Watching for changes in an object</h2>
-
-<div class="row">
-    <div class="col-md-6">
-
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">Status</h3>
-            </div>
-            <div class="panel-body">
-                <ol>
-                    <li>Add, delete, or delete all messages.</li>
-                    <li>The console logs a message when the token's <code>OnChange</code> method is called, which triggers a callback on the change token.<img width="272" height="57" alt="Console showing a message that the messages have changed and the type of change that occurred on 10/27/2017 at 9:32:24 AM" src="~/images/console-object-change.png"></li>
-                </ol>
-                <p>Last Message Change: <span class="label label-@ViewData["LastMessageChangeTypeStyle"]">@Model.LastMessageChangeType</span></p>
-            </div>
-        </div>
-
-    </div>
-</div>
-
-<div class="row">
-    <div class="col-md-6">
-
-        <form method="post">
-            <div class="panel panel-default">
-                <div class="panel-heading">
-                    <h3 class="panel-title">Add a message</h3>
-                </div>
-                <div class="panel-body">
-                    <div class="form-group">
-                        <label asp-for="Message.Text"></label>
-                        <input asp-for="Message.Text" class="form-control">
-                    </div>
-                    <div class="form-group">
-                        <button type="submit" asp-page-handler="AddMessage" class="btn btn-default">Add Message</button>
-                    </div>
-                </div>
-            </div>
-        </form>
-
-    </div>
-</div>
-
-<div class="row">
-    <div class="col-md-6">
-
-        <form method="post">
-            <div class="panel panel-default">
-                <div class="panel-heading clearfix">
-                    <button type="submit" asp-page-handler="DeleteAllMessages" class="pull-right btn btn-danger">Clear All</button>
-                    <h3 class="panel-title" style="line-height:2.1">Messages</h3>
-                </div>
-                <div class="panel-body">
-                    <ol>
-                        @foreach (var message in Model.Messages)
-                        {
-                            <li>
-                                <button type="submit" class="btn btn-danger" style="margin-bottom:2px" asp-page-handler="DeleteMessage" asp-route-id="@message.Id">Delete</button>
-                                @message.Text
-                            </li>
-                        }
-                    </ol>
-                </div>
-            </div>
-        </form>
-
-    </div>
-</div>
