+++ aspnetcore/fundamentals/primitives/change-tokens/sample/Pages/_Layout.cshtml
@@ -32,9 +32,6 @@
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
                     <li><a asp-page="/Index">Home</a></li>
-                    <li><a asp-page="/ConfigWatching">Config changes</a></li>
-                    <li><a asp-page="/FileWatching">File changes</a></li>
-                    <li><a asp-page="/ObjectWatching">Object changes</a></li>
                 </ul>
             </div>
         </div>
