+++ aspnetcore/fundamentals/primitives/change-tokens/sample/Pages/_ViewImports.cshtml
@@ -1,4 +1,3 @@
 @using ChangeTokenSample
-@using ChangeTokenSample.Data
 @namespace ChangeTokenSample.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
