+++ aspnetcore/fundamentals/primitives/change-tokens/sample/Utilities/Utilities.cs
@@ -1,4 +1,6 @@
+using System;
 using System.IO;
+using System.Threading;
 
 namespace ChangeTokenSample.Utilities
 {
@@ -7,16 +9,46 @@ namespace ChangeTokenSample.Utilities
         #region snippet1
         public static byte[] ComputeHash(string filePath)
         {
-            if (File.Exists(filePath))
+            var runCount = 1;
+
+            while(runCount < 4)
             {
-                using (var fs = File.OpenRead(filePath))
+                try
                 {
-                    return System.Security.Cryptography.SHA1.Create().ComputeHash(fs);
+                    if (File.Exists(filePath))
+                    {
+                        using (var fs = File.OpenRead(filePath))
+                        {
+                            return System.Security.Cryptography.SHA1.Create().ComputeHash(fs);
+                        }
+                    }
+                }
+                catch (IOException ex)
+                {
+                    if (runCount == 3 || ex.HResult != -2147024864)
+                    {
+                        throw;
+                    }
+                    else
+                    {
+                        Thread.Sleep(TimeSpan.FromSeconds(Math.Pow(2, runCount)));
+                        runCount++;
+                    }
                 }
             }
 
             return new byte[20];
         }
         #endregion
+
+        public static void WriteConsole(string s)
+        {
+            var consoleString = $"{DateTime.Now} {s}";
+            var l = consoleString.Length;
+
+            Console.WriteLine(new string('*', l));
+            Console.WriteLine(consoleString);
+            Console.WriteLine(new string('*', l));
+        }
     }
 }
